requirements.txt
aniso8601==8.1.0
appdirs==1.4.4
apturl==0.5.2
astroid==2.4.2
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
boto3==1.16.49
botocore==1.19.49
Brlapi==0.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.0.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Jsonpify==1.5.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
fpdf==1.7.2
future==0.18.2
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
httplib2==0.14.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
numpy==1.19.0
oauthlib==3.1.0
olefile==0.46
pandas==1.0.5
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.2.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.3.0
PyPDF2==1.26.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.4
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
s3transfer==0.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.26
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
flask_pymongo
